Extract composer diff review views
