+++ scripts/test-composer-identity-grid.js
@@ -34,6 +34,7 @@ const composerSeoStagingPath = resolve(here, '../assets/js/composer-seo-staging.
 const composerIndexTabsModelPath = resolve(here, '../assets/js/composer-index-tabs-model.js');
 const composerSiteModelPath = resolve(here, '../assets/js/composer-site-model.js');
 const composerDiffUiPath = resolve(here, '../assets/js/composer-diff-ui.js');
+const composerDiffReviewViewsPath = resolve(here, '../assets/js/composer-diff-review-views.js');
 const composerOrderDiffUiPath = resolve(here, '../assets/js/composer-order-diff-ui.js');
 const composerOrderPreviewPath = resolve(here, '../assets/js/composer-order-preview.js');
 const composerOrderVisualPath = resolve(here, '../assets/js/composer-order-visual.js');
@@ -167,6 +168,7 @@ const composerSeoStagingSource = readFileSync(composerSeoStagingPath, 'utf8');
 const composerIndexTabsModelSource = readFileSync(composerIndexTabsModelPath, 'utf8');
 const composerSiteModelSource = readFileSync(composerSiteModelPath, 'utf8');
 const composerDiffUiSource = readFileSync(composerDiffUiPath, 'utf8');
+const composerDiffReviewViewsSource = readFileSync(composerDiffReviewViewsPath, 'utf8');
 const composerOrderDiffUiSource = readFileSync(composerOrderDiffUiPath, 'utf8');
 const composerOrderPreviewSource = readFileSync(composerOrderPreviewPath, 'utf8');
 const composerOrderVisualSource = readFileSync(composerOrderVisualPath, 'utf8');
@@ -876,6 +878,12 @@ assert.match(
 
 assert.match(
   composerOrderDiffUiSource,
+  /from '\.\/composer-diff-review-views\.js'/,
+  'composer order diff UI should delegate overview and entries review rendering'
+);
+
+assert.match(
+  composerOrderDiffUiSource,
   /from '\.\/composer-order-visual\.js'/,
   'composer order diff UI should delegate visual connector and hover behavior'
 );
@@ -894,8 +902,14 @@ assert.doesNotMatch(
 
 assert.match(
   composerOrderDiffUiSource,
-  /export function createComposerOrderDiffUi\(options = \{\}\)[\s\S]*const setTimeoutRef = typeof options\.setTimeoutRef === 'function'[\s\S]*const requestAnimationFrameRef = typeof options\.requestAnimationFrameRef === 'function'[\s\S]*const addWindowListener = typeof options\.addWindowListener === 'function'[\s\S]*const addDocumentListener = typeof options\.addDocumentListener === 'function'[\s\S]*const consoleRef = options\.consoleRef \|\| null[\s\S]*const composerOrderVisual = createComposerOrderVisual\([\s\S]*const composerOrderPreview = createComposerOrderPreview\([\s\S]*function ensureComposerDiffModal\(\)[\s\S]*function closeComposerDiffModalForKind\(kind\)/,
-  'order diff UI boundary should own composer review modal while wiring visual and inline preview boundaries'
+  /export function createComposerOrderDiffUi\(options = \{\}\)[\s\S]*const setTimeoutRef = typeof options\.setTimeoutRef === 'function'[\s\S]*const requestAnimationFrameRef = typeof options\.requestAnimationFrameRef === 'function'[\s\S]*const addWindowListener = typeof options\.addWindowListener === 'function'[\s\S]*const addDocumentListener = typeof options\.addDocumentListener === 'function'[\s\S]*const consoleRef = options\.consoleRef \|\| null[\s\S]*const composerOrderVisual = createComposerOrderVisual\([\s\S]*const composerOrderPreview = createComposerOrderPreview\([\s\S]*const composerDiffReviewViews = createComposerDiffReviewViews\([\s\S]*function ensureComposerDiffModal\(\)[\s\S]*function closeComposerDiffModalForKind\(kind\)/,
+  'order diff UI boundary should own composer review modal shell while wiring review views, visual, and inline preview boundaries'
+);
+
+assert.match(
+  composerDiffReviewViewsSource,
+  /export function createComposerDiffReviewViews\(options = \{\}\)[\s\S]*function renderOverview\(target, diff\)[\s\S]*function describeEntrySnapshot\(kind, key, source\)[\s\S]*function buildEntryDetails\(kind, key, info, sectionType\)[\s\S]*function renderEntries\(target, kind, diff\)/,
+  'diff review views boundary should own overview and entries tab DOM rendering'
 );
 
 assert.match(
@@ -912,8 +926,8 @@ assert.match(
 
 assert.doesNotMatch(
   composerOrderDiffUiSource,
-  /const ORDER_LINE_COLORS|function getComposerOrderHoverContainer|function applyComposerOrderHover\(container, key\)|function bindComposerOrderHover\(element, key\)|function buildOrderDiffItem\(entry, side\)[\s\S]*item\.appendChild\(badgeEl\);|function ensureComposerOrderPreview\(kind\)|function updateComposerOrderPreview\(kind, options = \{\}\)|let composerOrderPreviewState|const composerOrderPreviewRelayoutTimers/,
-  'order diff UI should not re-own visual connector or inline preview internals'
+  /const ORDER_LINE_COLORS|function getComposerOrderHoverContainer|function applyComposerOrderHover\(container, key\)|function bindComposerOrderHover\(element, key\)|function buildOrderDiffItem\(entry, side\)[\s\S]*item\.appendChild\(badgeEl\);|function ensureComposerOrderPreview\(kind\)|function updateComposerOrderPreview\(kind, options = \{\}\)|let composerOrderPreviewState|const composerOrderPreviewRelayoutTimers|function renderOverview\(target, diff\)|function describeEntrySnapshot\(kind, key, source\)|function buildEntryDetails\(kind, key, info, sectionType\)|function renderEntries\(target, kind, diff\)/,
+  'order diff UI should not re-own visual connector, inline preview, or review-tab rendering internals'
 );
 
 assert.match(
@@ -925,6 +939,7 @@ assert.match(
 assert.doesNotMatch(
   [
     composerOrderDiffUiSource,
+    composerDiffReviewViewsSource,
     composerOrderPreviewSource,
     composerOrderVisualSource,
     composerEditorShellSource,
